makes specification for covariates and puts it into the single estimation main function
