+++ src/cascade/executor/covariate_data.py
@@ -1,0 +1,14 @@
+def add_covariate_data_to_observations_and_avgints(execution_context, data, epiviz_covariates):
+    """
+
+    Args:
+        execution_context: Execution context for db and file access.
+        data: The object that contains observations, average integrands
+            and raw covariate data.
+        epiviz_covariates List[EpiVizCovariate]: The specification for the
+            covariates.
+
+    Returns:
+
+    """
+    pass
