+++ tests/input_data/db/test_study_covariates.py
@@ -6,7 +6,10 @@ import pandas as pd
 import pytest
 
 from cascade.input_data import InputDataError
-from cascade.input_data.configuration.construct_study import _normalize_covariate_data
+from cascade.input_data.configuration.construct_study import (
+    _normalize_covariate_data, get_study_covariates
+)
+from cascade.testing_utilities import make_execution_context
 
 
 @pytest.fixture
@@ -62,9 +65,10 @@ def test_id_disagrees(basic_bundle):
     """
     id_to_name = {102: "smoking", 64: "love_polka"}
     not_in_bundle_index = 20
-    covs_in = pd.DataFrame({"study_covariate_id": [64, 102, 64, 64, 102],
-                  "bundle_id": [77, 77, 77, 77, 77],
-                  "seq": [4, 4, 6, 8, not_in_bundle_index]})
+    covs_in = pd.DataFrame(
+        {"study_covariate_id": [64, 102, 64, 64, 102],
+         "bundle_id": [77, 77, 77, 77, 77],
+         "seq": [4, 4, 6, 8, not_in_bundle_index]})
     with pytest.raises(InputDataError):
         _normalize_covariate_data(basic_bundle, covs_in, id_to_name)
 
@@ -74,9 +78,17 @@ def test_no_covariates(basic_bundle):
     Given covariate data and a bundle, create Series that correspond to
     the covariate values for the bundle bundle.
     """
-    cov_in = pd.DataFrame({"study_covariate_id": [],
-                         "bundle_id": [],
-                         "seq": []})
+    cov_in = pd.DataFrame(
+        {"study_covariate_id": [],
+         "bundle_id": [],
+         "seq": []})
     covs = pd.DataFrame({"covariate_sequence_number": basic_bundle.seq})
     normalized = _normalize_covariate_data(basic_bundle, cov_in, {})
     pd.testing.assert_frame_equal(normalized, covs)
+
+
+def test_get_study_covariates(ihme):
+    ec = make_execution_context()
+    covariates = get_study_covariates(
+        ec, bundle_id=4325, model_version_id=268082, tier=2)
+    assert set(covariates.columns.tolist()) == {"bundle_id", "seq", "study_covariate_id"}
